# Backend/src/api/openapi.yaml
openapi: 3.0.0
info:
  title: Vehicle Inventory
  description: "Das Vehicle Inventory beinhaltet diverse Fahrzeuge verschiedener Marken.\
    \ Die Daten, welche im Backend zur Verfügung gestellt werden, können mittels \
    \ API Abfrage von verschiedenen Frontends optisch aufbereitet werden. \
    \ Die Informationen beinhalten Marke, Modell und Leistung (PS)."
  contact:
    name: Philipp Lehnet 
    url: https://www.bl-itsolutions.de
    email: lehnet@bl-itsolutions.de
  license:
    name: MIT
    url: https://mit-license.org
  version: "1.0"
servers:
- url: http://localhost:3000
...
paths:
...
  /vehicle/{id}:
    get:
      tags:
      - Vehicle
      summary: fetch vehicle
      description: Fetches a single vehicle from the database.
      operationId: get-vehicle
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
        "404":
          description: Not Found
    put:
      tags:
      - Overwrite vehicle
      summary: ""
      description: "Overwrites a single vehicle in the database. You can edit all \
        \ attributes except the ID."
      operationId: update-vehicle
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Vehicle'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
        "404":
          description: Not Found
    patch:
      tags:
      - Change single attributes
      summary: Change single attributes
      description: "Changes single attributes of existing vehicles.\
        \ You are allowed to change any attribute except of the ID."
      operationId: patch-vehicle
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Vehicle'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
        "404":
          description: Not Found
    delete:
      tags:
      - Vehicle
      summary: Delete vehicle
      description: Deletes a single vehicle from the database.
      operationId: delete-vehicle
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
...
components:
  schemas:
...
    HATEOAS_Resource:
      title: HATEOAS-Links of a Ressource
      type: object
      properties:
        read:
          $ref: '#/components/schemas/HATEOAS_Link'
        update:
          $ref: '#/components/schemas/HATEOAS_Link'
        patch:
          $ref: '#/components/schemas/HATEOAS_Link'
        delete:
          $ref: '#/components/schemas/HATEOAS_Link'
    HATEOAS_Link:
      title: HATEOAS-Link
      type: object
      properties:
        url:
          type: string
        method:
          type: string
        query_params:
          type: array
          items:
            type: string
    Vehicle:
      title: Vehicle
      type: object
      properties:
        _links:
          $ref: '#/components/schemas/HATEOAS_Resource'
        _id:
          type: string
        vendor:
          type: string
        model:
          type: string
        power:
          type: string
      required:
        - vendor
        - model
      example:
        _id: "621ff9a936997df5baba746e"
        vendor: "Audi"
        model: "TT RS Coupe"
        power: "249 PS"
...
  securitySchemes: {}